Move ColonyDomains into ColonyMeta etc
