--- src/modules/dashboard/components/ColonyHome/ColonyDomains/ColonyDomains.css
@@ -1,22 +1,0 @@
-.filterItem {
-  composes: inlineEllipsis from '~styles/text.css';
-  margin: 0 0 20px 10px;
-  padding: 0;
-  max-width: 23ch;
-  border: none;
-  background: none;
-  font-weight: var(--weight-semi);
-  color: var(--text);
-  cursor: pointer;
-  outline: none;
-}
-
-.filterItem:hover {
-  color: var(--primary);
-  opacity: 0.8;
-}
-
-.filterItemActive {
-  composes: filterItem;
-  color: var(--primary);
-}
