--- src/modules/dashboard/components/ColonyHome/ColonyDomains/ColonyDomains.css.d.ts
@@ -1,2 +1,0 @@
-export const filterItem: string;
-export const filterItemActive: string;
