--- src/modules/dashboard/components/ColonyHome/ColonyDomains/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ColonyDomains';
