+++ src/modules/dashboard/components/ColonyHome/ColonyHome.css.d.ts
@@ -2,7 +2,6 @@ export const wideButton: string;
 export const main: string;
 export const colonyInfo: string;
 export const interactiveBar: string;
-export const domainContainer: string;
 export const metaContainer: string;
 export const content: string;
 export const noTasks: string;
