+++ src/modules/dashboard/components/ColonyHome/ColonyMeta/ColonyMeta.css.d.ts
@@ -1,14 +1,12 @@
 export const main: string;
 export const avatar: string;
 export const description: string;
-export const dynamicSection: string;
-export const link: string;
 export const editColony: string;
-export const avatarWrapper: string;
-export const userAvatar: string;
 export const colonyAvatarAndName: string;
 export const headingAndSettings: string;
-export const dynamicTextSection: string;
 export const simpleLinkWebsite: string;
 export const simpleLinkGuideline: string;
 export const spinnerContainer: string;
+export const domainContainer: string;
+export const filterItem: string;
+export const filterItemActive: string;
