+++ src/utils/strings/index.ts
@@ -22,7 +22,7 @@ export const rmLineBreaks = (str: string) => str.replace(/(\r\n|\n|\r)/gm, ' ');
  * @return {string} based on maxCharLength either the cut down string or the original one
  */
 export const multiLineTextEllipsis = (
-  string: string | null,
+  string: string,
   maxCharLength: number,
 ) => {
   if (string && string.length > maxCharLength) {
